update deckOfCards.js
Replaced random access array from the first test of iterations with a static method. SinglyLinkedList.
Obviously 1mill iterations is no longer a possibility however that was far beyond the performance I would hope for.
Reducing the iterations to 100*100 and the compilation time is almost instant. This example builds 52000 decks of cards under on name space.